+++ src/components/Callout/Callout.stories.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { createStoryFromTemplate, FictoanStory } from "../../utils/storyUtils";
+import { Callout } from "./Callout";
+
+const Template: FictoanStory<typeof Callout> = (args) => <Callout {...args} />;
+Template.displayName = Callout.displayName;
+
+export const Default: FictoanStory<typeof Callout> = createStoryFromTemplate(Template);
+Default.args = {
+    children: "Hello!"
+};