+++ CHANGELOG.md
@@ -10,8 +10,11 @@
 - RadioGroup and Select now uses BaseInputComponent
 - Add basic Tooltip component
 - Add better and consistent help & error text for each form field; Make InputLabel atomic
+- Add Callout component
+- Add error text for FormWrapper which uses Callout component
 
 **BREAKING CHANGES**
+- FormItemJointGroup removed
 - `helpText` is now part of inputField.default instead of inputField.onFocus inside theme
 - `helpText` inside inputField.isInvalid has been renamed to `errorText`